code cleanup CODBlackOps.asl
Removed `vars.doneMaps` (and all functions pertaining to it), as a player is unable to go to a previous map in this game. Removed timing message (can be reenabled) and instead sets timing method to game time whenever the user loads the script.
Removed useless brackets, set `map` string length to 14 (since the longest map name in `missions` is "underwaterbase").